requirements.txt
Flask==3.0.0
flask-cors==4.0.0
pdfplumber==0.10.3
google-generativeai==0.3.2
pydantic==2.5.3
python-dotenv==1.0.0
Werkzeug==3.0.1
gunicorn==21.2.0

